Move Home to Components and make uppercase the first letter of some folders and components
