+++ src/components/Home/HomeContainer.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react'
-import HeaderContainer from '../../components/Header/HeaderContainer'
+import HeaderContainer from '../Header/HeaderContainer'
 import Home from './Home'
 import httpService from '../../services/httpService'
